bsp(esp32_s3_eye): Regenerated by BSP Generator + moved to esp_video with camera
