+++ .github/ci/update_supported_boards.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# SPDX-FileCopyrightText: 2025 Espressif Systems (Shanghai) CO LTD
+# SPDX-FileCopyrightText: 2025-2026 Espressif Systems (Shanghai) CO LTD
 # SPDX-License-Identifier: Apache-2.0
 
 import re
@@ -77,51 +77,56 @@ def extract_board_info(board_path):
             if match:
                 features.add(match.group(1))
 
-        # Check for specific driver/controller names in C file
-        c_file = next(board_path.glob(f"{board_name}.c"), None)
-        # Some BSPs has main C file in src/ folder
-        if not c_file or not c_file.exists():
-            c_file = next(board_path.glob(f"src/{board_name}.c"), None)
-        if c_file and c_file.exists():
-            with open(c_file, "r", encoding="utf-8") as f:
-                c_code = f.read()
+        boardname_src_file = (board_name + '.c')
+        c_files = [boardname_src_file, "bsp_display.c", "bsp_audio.c"]
+        for file in c_files:
+            # Check for specific driver/controller names in C file
+            c_file = next(board_path.glob(file), None)
+            # Some BSPs has main C file in src/ folder
+            if not c_file or not c_file.exists():
+                c_file = next(board_path.glob(f"src/{file}"), None)
+            if c_file and c_file.exists():
+                with open(c_file, "r", encoding="utf-8") as f:
+                    c_code = f.read()
 
-                # DISPLAY: find all esp_lcd_new_panel_* excluding io_*
-                display_matches = re.findall(r"esp_lcd_new_panel_([a-z0-9_]+)\s*\(", c_code)
-                display_types = [m for m in display_matches if not m.startswith("io_")]
+                    if file == boardname_src_file or file == "bsp_display.c":
+                        # DISPLAY: find all esp_lcd_new_panel_* excluding io_*
+                        display_matches = re.findall(r"esp_lcd_new_panel_([a-z0-9_]+)\s*\(", c_code)
+                        display_types = [m for m in display_matches if not m.startswith("io_")]
 
-                # TOUCH: esp_lcd_touch_new_i2c_*
-                touch_types = re.findall(r"esp_lcd_touch_new_i2c_([a-z0-9_]+)\s*\(", c_code)
+                        # TOUCH: esp_lcd_touch_new_i2c_*
+                        touch_types = re.findall(r"esp_lcd_touch_new_i2c_([a-z0-9_]+)\s*\(", c_code)
 
-                # CAMERA: find camera sensors
-                if "CAMERA" in features:
-                    match = re.search(r"Supported camera sensors:\s*(.*)", content)
-                    if match:
-                        cam_sensor = re.findall(r"[A-Z0-9_]+", match.group(1))
+                    if file == boardname_src_file or file == "bsp_audio.c":
+                        # AUDIO_SPEAKER
+                        if "AUDIO_SPEAKER" in features:
+                            speaker_code_section = extract_function_body(c_code, 'bsp_audio_codec_speaker_init')
+                            if speaker_code_section is not None:
+                                matches = re.findall(r"([a-z0-9_]+)_codec_new", speaker_code_section)
+                                speaker_codec = [m for m in matches if not m.startswith("audio")]
+                            if not speaker_codec:
+                                speaker_code_section = extract_function_body(c_code, 'bsp_audio_codec_init')
+                                if speaker_code_section is not None:
+                                    matches = re.findall(r"([a-z0-9_]+)_codec_new", speaker_code_section)
+                                    speaker_codec = [m for m in matches if not m.startswith("audio")]
 
-                # AUDIO_SPEAKER
-                if "AUDIO_SPEAKER" in features:
-                    speaker_code_section = extract_function_body(c_code, 'bsp_audio_codec_speaker_init')
-                    if speaker_code_section is not None:
-                        matches = re.findall(r"([a-z0-9_]+)_codec_new", speaker_code_section)
-                        speaker_codec = [m for m in matches if not m.startswith("audio")]
-                    if not speaker_codec:
-                        speaker_code_section = extract_function_body(c_code, 'bsp_audio_codec_init')
-                        if speaker_code_section is not None:
-                            matches = re.findall(r"([a-z0-9_]+)_codec_new", speaker_code_section)
-                            speaker_codec = [m for m in matches if not m.startswith("audio")]
+                        # AUDIO_MIC
+                        if "AUDIO_MIC" in features:
+                            mic_code_section = extract_function_body(c_code, 'bsp_audio_codec_microphone_init')
+                            if mic_code_section is not None:
+                                matches = re.findall(r"([a-z0-9_]+)_codec_new", mic_code_section)
+                                mic_codec = [m for m in matches if not m.startswith("audio")]
+                            if not mic_codec:
+                                mic_code_section = extract_function_body(c_code, 'bsp_audio_codec_init')
+                                if mic_code_section is not None:
+                                    matches = re.findall(r"([a-z0-9_]+)_codec_new", mic_code_section)
+                                    mic_codec = [m for m in matches if not m.startswith("audio")]
 
-                # AUDIO_MIC
-                if "AUDIO_MIC" in features:
-                    mic_code_section = extract_function_body(c_code, 'bsp_audio_codec_microphone_init')
-                    if mic_code_section is not None:
-                        matches = re.findall(r"([a-z0-9_]+)_codec_new", mic_code_section)
-                        mic_codec = [m for m in matches if not m.startswith("audio")]
-                    if not mic_codec:
-                        mic_code_section = extract_function_body(c_code, 'bsp_audio_codec_init')
-                        if mic_code_section is not None:
-                            matches = re.findall(r"([a-z0-9_]+)_codec_new", mic_code_section)
-                            mic_codec = [m for m in matches if not m.startswith("audio")]
+            # CAMERA: find camera sensors
+            if "CAMERA" in features:
+                match = re.search(r"Supported camera sensors:\s*(.*)", content)
+                if match:
+                    cam_sensor = re.findall(r"[A-Z0-9_]+", match.group(1))
 
     # Extract SoC from YML
     with open(yml_file, "r", encoding="utf-8") as f:
