+++ bsp/esp32_p4_eye/idf_component.yml
@@ -41,5 +41,5 @@ examples:
   - path: ../../examples/display
   - path: ../../examples/display_rotation
   - path: ../../examples/display_lvgl_demos
-  - path: ../../examples/display_camera_csi
+  - path: ../../examples/display_camera_video
