+++ bsp/esp32_p4_eye/include/bsp/esp32_p4_eye.h
@@ -565,7 +565,7 @@ void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation);
  *
  * Camera interface
  * Supported camera sensors: OV2710
- * More information in display_camera_csi example
+ * More information in display_camera_video example
  *
  **************************************************************************************************/
