+++ bsp/esp32_p4_function_ev_board/idf_component.yml
@@ -50,4 +50,4 @@ examples:
   - path: ../../examples/display_lvgl_benchmark
   - path: ../../examples/display_usb_hid
   - path: ../../examples/display_sdcard
-  - path: ../../examples/display_camera_csi
+  - path: ../../examples/display_camera_video
