+++ bsp/esp32_s3_eye/Kconfig
@@ -1,18 +1,17 @@
-menu "Board Support Package"
+menu "Board Support Package (ESP32-S3-EYE)"
 
     config BSP_ERROR_CHECK
         bool "Enable error check in BSP"
         default y
         help
             Error check assert the application before returning the error code.
-
     menu "I2C"
         config BSP_I2C_NUM
             int "I2C peripheral index"
             default 1
             range 0 1
             help
-                ESP32-S3 has two I2C peripherals, pick the one you want to use.
+                Pick the I2C you want to use.
 
         config BSP_I2C_FAST_MODE
             bool "Enable I2C fast mode"
@@ -25,87 +24,84 @@ menu "Board Support Package"
             default 400000 if BSP_I2C_FAST_MODE
             default 100000
     endmenu
+    menu "Storage"
+        menu "SPIFFS - Virtual File System"
+            config BSP_SPIFFS_FORMAT_ON_MOUNT_FAIL
+                bool "Format SPIFFS if mounting fails"
+                default n
+                help
+                    Format SPIFFS if it fails to mount the filesystem.
 
-    menu "SPIFFS - Virtual File System"
-        config BSP_SPIFFS_FORMAT_ON_MOUNT_FAIL
-            bool "Format SPIFFS if mounting fails"
-            default n
-            help
-                Format SPIFFS if it fails to mount the filesystem.
+            config BSP_SPIFFS_MOUNT_POINT
+                string "SPIFFS mount point"
+                default "/spiffs"
+                help
+                    Mount point of SPIFFS in the Virtual File System.
 
-        config BSP_SPIFFS_MOUNT_POINT
-            string "SPIFFS mount point"
-            default "/spiffs"
-            help
-                Mount point of SPIFFS in the Virtual File System.
+            config BSP_SPIFFS_PARTITION_LABEL
+                string "Partition label of SPIFFS"
+                default "storage"
+                help
+                    Partition label which stores SPIFFS.
 
-        config BSP_SPIFFS_PARTITION_LABEL
-            string "Partition label of SPIFFS"
-            default "storage"
-            help
-                Partition label which stores SPIFFS.
+            config BSP_SPIFFS_MAX_FILES
+                int "Max files supported for SPIFFS VFS"
+                default 5
+                help
+                    Supported max files for SPIFFS in the Virtual File System.
+        endmenu
+        menu "SD card - Virtual File System"
+            config BSP_SD_FORMAT_ON_MOUNT_FAIL
+                bool "Format SD card if mounting fails"
+                default n
+                help
+                    The SDMMC host will format (FAT) the SD card if it fails to mount the filesystem.
 
-        config BSP_SPIFFS_MAX_FILES
-            int "Maximum files that could be open at the same time"
-            default 2
-            help
-                Supported max files for SPIFFS in the Virtual File System.
-    endmenu
+            config BSP_SD_MOUNT_POINT
+                string "SD card mount point"
+                default "/sdcard"
+                help
+                    Mount point of the SD card in the Virtual File System
 
-    menu "uSD card - Virtual File System"
-        config BSP_SD_FORMAT_ON_MOUNT_FAIL
-            bool "Format uSD card if mounting fails"
-            default n
-            help
-                The SDMMC host will format (FAT) the uSD card if it fails to mount the filesystem.
-
-        config BSP_SD_MOUNT_POINT
-            string "uSD card mount point"
-            default "/sdcard"
-            help
-                Mount point of the uSD card in the Virtual File System
-
+        endmenu
     endmenu
     menu "Display"
-        config BSP_DISPLAY_LVGL_TASK_PRIORITY
-            int "LVGL task priority"
-            default 2
-            help
-                The Board Support Package will create a task that will periodically handle LVGL operation in lv_timer_handler().
-
-        config BSP_DISPLAY_LVGL_TICK
-            int "LVGL tick period"
-            default 5
-            range 1 100
-            help
-                Period of LVGL tick timer.
-
-        config BSP_DISPLAY_LVGL_MAX_SLEEP
-            int "LVGL maximum sleep time in ms"
-            default 1
-            range 1 500
-            help
-                Maximum time for task sleep in ms.
-        
-        config BSP_DISPLAY_LVGL_BUFFER_IN_PSRAM
-            bool "Allocate LVGL buffer in PSRAM"
-            default n
-            help
-                LVGL buffer will be allocated in PSRAM instead of internal RAM with DMA.
-
         config BSP_DISPLAY_BRIGHTNESS_LEDC_CH
-            int "LEDC channel index"
-            default 1
-            range 0 7
-            help
-                LEDC channel is used to generate PWM signal that controls display brightness.
-                Set LEDC index that should be used.
-    endmenu
-
-    config BSP_I2S_NUM
-        int "I2S peripheral index"
+        int "LEDC channel index"
         default 1
-        range 0 1
+        range 0 7
         help
-            ESP32-S3 has two I2S peripherals, pick the one you want to use.
+            LEDC channel is used to generate PWM signal that controls display brightness.
+            Set LEDC index that should be used.
+                
+        config BSP_LCD_DRAW_BUF_HEIGHT
+        int "LCD framebuf height"
+        default 100
+        range 10 240
+        help
+            Framebuf is used for lvgl rendering output.
+
+        config BSP_LCD_DRAW_BUF_DOUBLE
+        bool "LCD double framebuf"
+        default y
+        help
+            Whether to enable double framebuf.
+    endmenu
+    menu "Camera"
+        config BSP_CAMERA_XCLK_LEDC_CH
+        int "LEDC channel index"
+        default 2
+        range 0 7
+        help
+            LEDC channel is used to generate PWM signal that controls display brightness.
+            Set LEDC index that should be used.
+    endmenu
+    menu "Audio"
+        config BSP_I2S_NUM
+            int "I2S peripheral index"
+            default 1
+            range 0 1
+            help
+                Pick the I2S you want to use.
+    endmenu
 endmenu
