--- bsp/esp32_s3_eye/esp32_s3_eye.c
@@ -1,561 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022-2026 Espressif Systems (Shanghai) CO LTD
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#include <stdio.h>
-#include <string.h>
-#include "esp_vfs_fat.h"
-#include "esp_lcd_panel_io.h"
-#include "esp_lcd_panel_vendor.h"
-#include "esp_lcd_panel_ops.h"
-#include "esp_log.h"
-#include "esp_check.h"
-#include "esp_vfs_fat.h"
-#include "esp_spiffs.h"
-
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-
-#include "driver/spi_master.h"
-#include "driver/ledc.h"
-#include "driver/i2c_master.h"
-
-#include "bsp/esp32_s3_eye.h"
-#include "bsp_err_check.h"
-#include "bsp/display.h"
-#include "button_gpio.h"
-#include "button_adc.h"
-
-static const char *TAG = "S3-EYE";
-
-/**
- * @brief I2C handle for BSP usage
- *
- * In IDF v5.4 you can call i2c_master_get_bus_handle(BSP_I2C_NUM, i2c_master_bus_handle_t *ret_handle)
- * from #include "esp_private/i2c_platform.h" to get this handle
- *
- * For IDF 5.2 and 5.3 you must call bsp_i2c_get_handle()
- */
-static i2c_master_bus_handle_t i2c_handle = NULL;
-static bool i2c_initialized = false;
-static adc_oneshot_unit_handle_t bsp_adc_handle = NULL;
-static sdmmc_card_t *bsp_sdcard = NULL;    // Global uSD card handler
-
-typedef enum {
-    BSP_BUTTON_TYPE_GPIO,
-    BSP_BUTTON_TYPE_ADC
-} bsp_button_type_t;
-
-typedef struct {
-    bsp_button_type_t type;
-    union {
-        button_gpio_config_t gpio;
-        button_adc_config_t  adc;
-    } cfg;
-} bsp_button_config_t;
-
-static const bsp_button_config_t bsp_button_config[BSP_BUTTON_NUM] = {
-    {
-        .type = BSP_BUTTON_TYPE_ADC,
-        .cfg.adc = {
-            .adc_handle = &bsp_adc_handle,
-            .adc_channel = ADC_CHANNEL_0, // ADC1 channel 0 is GPIO1
-            .button_index = BSP_BUTTON_MENU,
-            .min = 2310, // middle is 2410mV
-            .max = 2510
-        }
-    },
-    {
-        .type = BSP_BUTTON_TYPE_ADC,
-        .cfg.adc = {
-            .adc_handle = &bsp_adc_handle,
-            .adc_channel = ADC_CHANNEL_0, // ADC1 channel 0 is GPIO1
-            .button_index = BSP_BUTTON_PLAY,
-            .min = 1880, // middle is 1980mV
-            .max = 2080
-        }
-    },
-    {
-        .type = BSP_BUTTON_TYPE_ADC,
-        .cfg.adc = {
-            .adc_handle = &bsp_adc_handle,
-            .adc_channel = ADC_CHANNEL_0, // ADC1 channel 0 is GPIO1
-            .button_index = BSP_BUTTON_DOWN,
-            .min = 720, // middle is 820mV
-            .max = 920
-        }
-    },
-    {
-        .type = BSP_BUTTON_TYPE_ADC,
-        .cfg.adc = {
-            .adc_handle = &bsp_adc_handle,
-            .adc_channel = ADC_CHANNEL_0, // ADC1 channel 0 is GPIO1
-            .button_index = BSP_BUTTON_UP,
-            .min = 280, // middle is 380mV
-            .max = 480
-        }
-    },
-    {
-        .type = BSP_BUTTON_TYPE_GPIO,
-        .cfg.gpio = {
-            .active_level = 0,
-            .gpio_num = BSP_BUTTON_BOOT_IO
-        }
-    }
-};
-
-esp_err_t bsp_i2c_init(void)
-{
-    /* I2C was initialized before */
-    if (i2c_initialized) {
-        return ESP_OK;
-    }
-
-    const i2c_master_bus_config_t i2c_config = {
-        .i2c_port = BSP_I2C_NUM,
-        .sda_io_num = BSP_I2C_SDA,
-        .scl_io_num = BSP_I2C_SCL,
-        .clk_source = I2C_CLK_SRC_DEFAULT,
-        .flags.enable_internal_pullup = true,
-    };
-    BSP_ERROR_CHECK_RETURN_ERR(i2c_new_master_bus(&i2c_config, &i2c_handle));
-
-    i2c_initialized = true;
-    return ESP_OK;
-}
-
-esp_err_t bsp_i2c_deinit(void)
-{
-    BSP_ERROR_CHECK_RETURN_ERR(i2c_del_master_bus(i2c_handle));
-    i2c_initialized = false;
-    return ESP_OK;
-}
-
-i2c_master_bus_handle_t bsp_i2c_get_handle(void)
-{
-    bsp_i2c_init();
-    return i2c_handle;
-}
-
-esp_err_t bsp_spiffs_mount(void)
-{
-    esp_vfs_spiffs_conf_t conf = {
-        .base_path = CONFIG_BSP_SPIFFS_MOUNT_POINT,
-        .partition_label = CONFIG_BSP_SPIFFS_PARTITION_LABEL,
-        .max_files = CONFIG_BSP_SPIFFS_MAX_FILES,
-#ifdef CONFIG_BSP_SPIFFS_FORMAT_ON_MOUNT_FAIL
-        .format_if_mount_failed = true,
-#else
-        .format_if_mount_failed = false,
-#endif
-    };
-
-    esp_err_t ret_val = esp_vfs_spiffs_register(&conf);
-
-    BSP_ERROR_CHECK_RETURN_ERR(ret_val);
-
-    size_t total = 0, used = 0;
-    ret_val = esp_spiffs_info(conf.partition_label, &total, &used);
-    if (ret_val != ESP_OK) {
-        ESP_LOGE(TAG, "Failed to get SPIFFS partition information (%s)", esp_err_to_name(ret_val));
-    } else {
-        ESP_LOGI(TAG, "Partition size: total: %d, used: %d", total, used);
-    }
-
-    return ret_val;
-}
-
-esp_err_t bsp_spiffs_unmount(void)
-{
-    return esp_vfs_spiffs_unregister(CONFIG_BSP_SPIFFS_PARTITION_LABEL);
-}
-
-sdmmc_card_t *bsp_sdcard_get_handle(void)
-{
-    return bsp_sdcard;
-}
-
-void bsp_sdcard_get_sdmmc_host(const int slot, sdmmc_host_t *config)
-{
-    assert(config);
-
-    sdmmc_host_t host_config = SDMMC_HOST_DEFAULT();
-
-    memcpy(config, &host_config, sizeof(sdmmc_host_t));
-}
-
-void bsp_sdcard_get_sdspi_host(const int slot, sdmmc_host_t *config)
-{
-    assert(config);
-    memset(config, 0, sizeof(sdmmc_host_t));
-    ESP_LOGE(TAG, "SD card SPI mode is not supported by HW!");
-}
-
-void bsp_sdcard_sdmmc_get_slot(const int slot, sdmmc_slot_config_t *config)
-{
-    assert(config);
-    memset(config, 0, sizeof(sdmmc_slot_config_t));
-
-    config->clk = BSP_SD_CLK;
-    config->cmd = BSP_SD_CMD;
-    config->d0 = BSP_SD_D0;
-    config->d1 = GPIO_NUM_NC;
-    config->d2 = GPIO_NUM_NC;
-    config->d3 = GPIO_NUM_NC;
-    config->d4 = GPIO_NUM_NC;
-    config->d5 = GPIO_NUM_NC;
-    config->d6 = GPIO_NUM_NC;
-    config->d7 = GPIO_NUM_NC;
-    config->cd = SDMMC_SLOT_NO_CD;
-    config->wp = SDMMC_SLOT_NO_WP;
-    config->width = 1;
-    config->flags = SDMMC_SLOT_FLAG_INTERNAL_PULLUP;
-}
-
-void bsp_sdcard_sdspi_get_slot(const spi_host_device_t spi_host, sdspi_device_config_t *config)
-{
-    assert(config);
-    memset(config, 0, sizeof(sdspi_device_config_t));
-    ESP_LOGE(TAG, "SD card SPI mode is not supported by HW!");
-}
-
-esp_err_t bsp_sdcard_sdmmc_mount(bsp_sdcard_cfg_t *cfg)
-{
-    sdmmc_host_t sdhost = {0};
-    sdmmc_slot_config_t sdslot = {0};
-    const esp_vfs_fat_sdmmc_mount_config_t mount_config = {
-#ifdef CONFIG_BSP_SD_FORMAT_ON_MOUNT_FAIL
-        .format_if_mount_failed = true,
-#else
-        .format_if_mount_failed = false,
-#endif
-        .max_files = 5,
-        .allocation_unit_size = 16 * 1024
-    };
-    assert(cfg);
-
-    if (!cfg->mount) {
-        cfg->mount = &mount_config;
-    }
-
-    if (!cfg->host) {
-        bsp_sdcard_get_sdmmc_host(SDMMC_HOST_SLOT_0, &sdhost);
-        cfg->host = &sdhost;
-    }
-
-    if (!cfg->slot.sdmmc) {
-        bsp_sdcard_sdmmc_get_slot(SDMMC_HOST_SLOT_0, &sdslot);
-        cfg->slot.sdmmc = &sdslot;
-    }
-
-#if !CONFIG_FATFS_LONG_FILENAMES
-    ESP_LOGW(TAG, "Warning: Long filenames on SD card are disabled in menuconfig!");
-#endif
-
-    return esp_vfs_fat_sdmmc_mount(BSP_SD_MOUNT_POINT, cfg->host, cfg->slot.sdmmc, cfg->mount, &bsp_sdcard);
-}
-
-esp_err_t bsp_sdcard_sdspi_mount(bsp_sdcard_cfg_t *cfg)
-{
-    ESP_LOGE(TAG, "SD card SPI mode is not supported by HW!");
-    return ESP_ERR_NOT_SUPPORTED;
-}
-
-esp_err_t bsp_sdcard_mount(void)
-{
-    bsp_sdcard_cfg_t cfg = {0};
-    return bsp_sdcard_sdmmc_mount(&cfg);
-}
-
-esp_err_t bsp_sdcard_unmount(void)
-{
-    esp_err_t ret = ESP_OK;
-
-    ret |= esp_vfs_fat_sdcard_unmount(BSP_SD_MOUNT_POINT, bsp_sdcard);
-    bsp_sdcard = NULL;
-
-    return ret;
-}
-
-#define LCD_CMD_BITS         (8)
-#define LCD_PARAM_BITS       (8)
-#define LCD_LEDC_CH          (CONFIG_BSP_DISPLAY_BRIGHTNESS_LEDC_CH)
-#define LVGL_TICK_PERIOD_MS  (CONFIG_BSP_DISPLAY_LVGL_TICK)
-#define LVGL_MAX_SLEEP_MS    (CONFIG_BSP_DISPLAY_LVGL_MAX_SLEEP)
-
-esp_err_t bsp_display_brightness_init(void)
-{
-    // Setup LEDC peripheral for PWM backlight control
-    const ledc_channel_config_t LCD_backlight_channel = {
-        .gpio_num = BSP_LCD_BACKLIGHT,
-        .speed_mode = LEDC_LOW_SPEED_MODE,
-        .channel = LCD_LEDC_CH,
-        .intr_type = LEDC_INTR_DISABLE,
-        .timer_sel = 1,
-        .duty = 0,
-        .hpoint = 0,
-        .flags.output_invert = true
-    };
-    const ledc_timer_config_t LCD_backlight_timer = {
-        .speed_mode = LEDC_LOW_SPEED_MODE,
-        .duty_resolution = LEDC_TIMER_10_BIT,
-        .timer_num = 1,
-        .freq_hz = 5000,
-        .clk_cfg = LEDC_AUTO_CLK
-    };
-
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_config(&LCD_backlight_timer));
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_channel_config(&LCD_backlight_channel));
-
-    return ESP_OK;
-}
-
-esp_err_t bsp_display_brightness_set(int brightness_percent)
-{
-    if (brightness_percent > 100) {
-        brightness_percent = 100;
-    } else if (brightness_percent < 0) {
-        brightness_percent = 0;
-    }
-
-    ESP_LOGI(TAG, "Setting LCD backlight: %d%%", brightness_percent);
-    // LEDC resolution set to 10bits, thus: 100% = 1023
-    uint32_t duty_cycle = (1023 * brightness_percent) / 100;
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_set_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH, duty_cycle));
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_update_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH));
-
-    return ESP_OK;
-}
-
-esp_err_t bsp_display_backlight_off(void)
-{
-    return bsp_display_brightness_set(0);
-}
-
-esp_err_t bsp_display_backlight_on(void)
-{
-    return bsp_display_brightness_set(100);
-}
-
-esp_err_t bsp_display_new(const bsp_display_config_t *config, esp_lcd_panel_handle_t *ret_panel,
-                          esp_lcd_panel_io_handle_t *ret_io)
-{
-    esp_err_t ret = ESP_OK;
-    assert(config != NULL && config->max_transfer_sz > 0);
-
-    ESP_RETURN_ON_ERROR(bsp_display_brightness_init(), TAG, "Brightness init failed");
-
-    ESP_LOGD(TAG, "Initialize SPI bus");
-    const spi_bus_config_t buscfg = {
-        .sclk_io_num = BSP_LCD_SPI_CLK,
-        .mosi_io_num = BSP_LCD_SPI_MOSI,
-        .miso_io_num = GPIO_NUM_NC,
-        .quadwp_io_num = GPIO_NUM_NC,
-        .quadhd_io_num = GPIO_NUM_NC,
-        .max_transfer_sz = config->max_transfer_sz,
-    };
-    ESP_RETURN_ON_ERROR(spi_bus_initialize(BSP_LCD_SPI_NUM, &buscfg, SPI_DMA_CH_AUTO), TAG, "SPI init failed");
-
-    ESP_LOGD(TAG, "Install panel IO");
-    const esp_lcd_panel_io_spi_config_t io_config = {
-        .dc_gpio_num = BSP_LCD_DC,
-        .cs_gpio_num = BSP_LCD_SPI_CS,
-        .pclk_hz = BSP_LCD_PIXEL_CLOCK_HZ,
-        .lcd_cmd_bits = LCD_CMD_BITS,
-        .lcd_param_bits = LCD_PARAM_BITS,
-        .spi_mode = 2,
-        .trans_queue_depth = 10,
-    };
-    ESP_GOTO_ON_ERROR(esp_lcd_new_panel_io_spi((esp_lcd_spi_bus_handle_t)BSP_LCD_SPI_NUM, &io_config, ret_io), err, TAG,
-                      "New panel IO failed");
-
-    ESP_LOGD(TAG, "Install LCD driver");
-    const esp_lcd_panel_dev_config_t panel_config = {
-        .reset_gpio_num = BSP_LCD_RST,
-        .rgb_ele_order = BSP_LCD_COLOR_SPACE,
-        .bits_per_pixel = BSP_LCD_BITS_PER_PIXEL,
-    };
-    ESP_GOTO_ON_ERROR(esp_lcd_new_panel_st7789(*ret_io, &panel_config, ret_panel), err, TAG, "New panel failed");
-
-
-    esp_lcd_panel_reset(*ret_panel);
-    esp_lcd_panel_init(*ret_panel);
-    esp_lcd_panel_invert_color(*ret_panel, true);
-    return ret;
-
-err:
-    if (*ret_panel) {
-        esp_lcd_panel_del(*ret_panel);
-    }
-    if (*ret_io) {
-        esp_lcd_panel_io_del(*ret_io);
-    }
-    spi_bus_free(BSP_LCD_SPI_NUM);
-    return ret;
-}
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static lv_display_t *bsp_display_lcd_init(const bsp_display_cfg_t *cfg)
-{
-    assert(cfg != NULL);
-    esp_lcd_panel_io_handle_t io_handle = NULL;
-    esp_lcd_panel_handle_t panel_handle = NULL;
-    const bsp_display_config_t bsp_disp_cfg = {
-        .max_transfer_sz = BSP_LCD_DRAW_BUFF_SIZE * sizeof(uint16_t),
-    };
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_display_new(&bsp_disp_cfg, &panel_handle, &io_handle));
-
-    esp_lcd_panel_disp_on_off(panel_handle, true);
-
-    /* Add LCD screen */
-    ESP_LOGD(TAG, "Add LCD screen");
-    const lvgl_port_display_cfg_t disp_cfg = {
-        .io_handle = io_handle,
-        .panel_handle = panel_handle,
-        .buffer_size = cfg->buffer_size,
-        .double_buffer = cfg->double_buffer,
-        .hres = BSP_LCD_H_RES,
-        .vres = BSP_LCD_V_RES,
-        .monochrome = false,
-        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
-        .rotation = {
-            .swap_xy = false,
-            .mirror_x = false,
-            .mirror_y = false,
-        },
-        .flags = {
-            .buff_dma = cfg->flags.buff_dma,
-            .buff_spiram = cfg->flags.buff_spiram,
-#if LVGL_VERSION_MAJOR >= 9
-            .swap_bytes = (BSP_LCD_BIGENDIAN ? true : false),
-#endif
-        }
-    };
-
-    return lvgl_port_add_disp(&disp_cfg);
-}
-
-lv_display_t *bsp_display_start(void)
-{
-    bsp_display_cfg_t cfg = {
-        .lvgl_port_cfg = {
-            .task_priority = CONFIG_BSP_DISPLAY_LVGL_TASK_PRIORITY,
-            .task_stack = 6144,
-            .task_affinity = 1,
-            .timer_period_ms = LVGL_TICK_PERIOD_MS,
-            .task_max_sleep_ms = LVGL_MAX_SLEEP_MS,
-        },
-        .buffer_size = BSP_LCD_DRAW_BUFF_SIZE,
-        .double_buffer = BSP_LCD_DRAW_BUFF_DOUBLE,
-        .flags = {
-#if (CONFIG_BSP_DISPLAY_LVGL_BUFFER_IN_PSRAM && CONFIG_SPIRAM)
-            .buff_dma = false,
-            .buff_spiram = true,
-#else
-            .buff_dma = true,
-            .buff_spiram = false,
-#endif
-        }
-    };
-    return bsp_display_start_with_config(&cfg);
-}
-
-lv_display_t *bsp_display_start_with_config(const bsp_display_cfg_t *cfg)
-{
-    lv_display_t *disp;
-    assert(cfg != NULL);
-    BSP_ERROR_CHECK_RETURN_NULL(lvgl_port_init(&cfg->lvgl_port_cfg));
-    BSP_NULL_CHECK(disp = bsp_display_lcd_init(cfg), NULL);
-
-    return disp;
-}
-
-lv_indev_t *bsp_display_get_input_dev(void)
-{
-    return NULL;
-}
-
-bool bsp_display_lock(uint32_t timeout_ms)
-{
-    return lvgl_port_lock(timeout_ms);
-}
-
-void bsp_display_unlock(void)
-{
-    lvgl_port_unlock();
-}
-
-void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation)
-{
-    lv_disp_set_rotation(disp, rotation);
-}
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-
-esp_err_t bsp_leds_init(void)
-{
-    const gpio_config_t led_io_config = {
-        .pin_bit_mask = BIT64(BSP_LED_GREEN),
-        .mode = GPIO_MODE_OUTPUT_OD, // GPIO3 is connected directly to the LED (on board revision 2.1), so we use Open-drain here
-        .pull_up_en = GPIO_PULLUP_DISABLE,
-        .pull_down_en = GPIO_PULLDOWN_DISABLE,
-        .intr_type = GPIO_INTR_DISABLE
-    };
-    BSP_ERROR_CHECK_RETURN_ERR(gpio_config(&led_io_config));
-    return ESP_OK;
-}
-
-esp_err_t bsp_led_set(const bsp_led_t led_io, const bool on)
-{
-    BSP_ERROR_CHECK_RETURN_ERR(gpio_set_level((gpio_num_t) led_io, (uint32_t) on));
-    return ESP_OK;
-}
-
-esp_codec_dev_handle_t bsp_audio_codec_microphone_init(void)
-{
-    const audio_codec_data_if_t *i2s_data_if = bsp_audio_get_codec_itf();
-    if (i2s_data_if == NULL) {
-        /* Configure I2S peripheral and Power Amplifier */
-        BSP_ERROR_CHECK_RETURN_NULL(bsp_audio_init(NULL));
-        i2s_data_if = bsp_audio_get_codec_itf();
-    }
-    assert(i2s_data_if);
-
-    esp_codec_dev_cfg_t codec_dev_cfg = {
-        .dev_type = ESP_CODEC_DEV_TYPE_IN,
-        .codec_if = NULL,
-        .data_if = i2s_data_if,
-    };
-    return esp_codec_dev_new(&codec_dev_cfg);
-}
-
-esp_err_t bsp_iot_button_create(button_handle_t btn_array[], int *btn_cnt, int btn_array_size)
-{
-    esp_err_t ret = ESP_OK;
-    const button_config_t btn_config = {0};
-    if ((btn_array_size < BSP_BUTTON_NUM) ||
-            (btn_array == NULL)) {
-        return ESP_ERR_INVALID_ARG;
-    }
-    /* Initialize ADC and get ADC handle */
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_adc_initialize());
-    bsp_adc_handle = bsp_adc_get_handle();
-
-    if (btn_cnt) {
-        *btn_cnt = 0;
-    }
-    for (int i = 0; i < BSP_BUTTON_NUM; i++) {
-        if (bsp_button_config[i].type == BSP_BUTTON_TYPE_GPIO) {
-            ret |= iot_button_new_gpio_device(&btn_config, &bsp_button_config[i].cfg.gpio, &btn_array[i]);
-        } else if (bsp_button_config[i].type == BSP_BUTTON_TYPE_ADC) {
-            ret |= iot_button_new_adc_device(&btn_config, &bsp_button_config[i].cfg.adc, &btn_array[i]);
-        } else {
-            ESP_LOGW(TAG, "Unsupported button type!");
-        }
-        if (btn_cnt) {
-            (*btn_cnt)++;
-        }
-    }
-    return ret;
-}
