+++ bsp/esp32_s3_eye/include/bsp/display.h
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2023-2025 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2023-2026 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -17,8 +17,7 @@
 #pragma once
 #include "esp_lcd_types.h"
 
-/** @defgroup g04_display Display and Touch
- *  @brief Display BSP API
+/** \addtogroup g04_display
  *  @{
  */
 
@@ -34,7 +33,8 @@
 #define BSP_LCD_BITS_PER_PIXEL      (16)
 /* LCD display color space */
 #define BSP_LCD_COLOR_SPACE         (LCD_RGB_ELEMENT_ORDER_RGB)
-/* LCD display definition */
+
+/* LCD definition */
 #define BSP_LCD_H_RES              (240)
 #define BSP_LCD_V_RES              (240)
 
@@ -43,12 +43,22 @@ extern "C" {
 #endif
 
 /**
- * @brief BSP display configuration structure
+ * @brief BSP display low level configuration structure
  *
  */
 typedef struct {
     int max_transfer_sz;    /*!< Maximum transfer size, in bytes. */
 } bsp_display_config_t;
+
+/**
+ * @brief BSP display return handles
+ *
+ */
+typedef struct {
+    esp_lcd_panel_io_handle_t   io;            /*!< ESP LCD IO handle */
+    esp_lcd_panel_handle_t      panel;         /*!< ESP LCD panel (color) handle */
+    esp_lcd_panel_handle_t      control;       /*!< ESP LCD panel (control) handle */
+} bsp_lcd_handles_t;
 
 /**
  * @brief Create new display panel
@@ -71,10 +81,37 @@ typedef struct {
  * @param[out] ret_io    esp_lcd IO handle
  * @return
  *      - ESP_OK         On success
- *      - Else           esp_lcd failure
+ *      - Other errors from underlying esp_lcd driver
  */
 esp_err_t bsp_display_new(const bsp_display_config_t *config, esp_lcd_panel_handle_t *ret_panel,
                           esp_lcd_panel_io_handle_t *ret_io);
+
+/**
+ * @brief Create new display panel
+ *
+ * For maximum flexibility, this function performs only reset and initialization of the display.
+ * You must turn on the display explicitly by calling esp_lcd_panel_disp_on_off().
+ * The display's backlight is not turned on either. You can use bsp_display_backlight_on/off(),
+ * bsp_display_brightness_set() (on supported boards) or implement your own backlight control.
+ *
+ * If you want to free resources allocated by this function, you can use API:
+ *
+ * \code{.c}
+ * bsp_display_delete();
+ * \endcode
+ *
+ * @param[in]  config    display configuration
+ * @param[out] ret_handles all esp_lcd handles in one structure
+ * @return
+ *      - ESP_OK         On success
+ *      - Other errors from underlying esp_lcd driver
+ */
+esp_err_t bsp_display_new_with_handles(const bsp_display_config_t *config, bsp_lcd_handles_t *ret_handles);
+
+/**
+ * @brief Delete display panel
+ */
+void bsp_display_delete(void);
 
 /**
  * @brief Initialize display's brightness
@@ -86,6 +123,15 @@ esp_err_t bsp_display_new(const bsp_display_config_t *config, esp_lcd_panel_hand
  *      - ESP_ERR_INVALID_ARG   Parameter error
  */
 esp_err_t bsp_display_brightness_init(void);
+
+/**
+ * @brief Deinitialize display's brightness
+ *
+ * @return
+ *      - ESP_OK                On success
+ *      - ESP_ERR_INVALID_ARG   Parameter error
+ */
+esp_err_t bsp_display_brightness_deinit(void);
 
 /**
  * @brief Set display's brightness
