+++ bsp/esp32_s3_eye/src/bsp_adc.c
@@ -1,0 +1,30 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "bsp_err_check.h"
+#include "bsp/esp32_s3_eye.h"
+static adc_oneshot_unit_handle_t bsp_adc_handle = NULL;
+
+esp_err_t bsp_adc_initialize(void)
+{
+    /* ADC was initialized before */
+    if (bsp_adc_handle != NULL) {
+        return ESP_OK;
+    }
+
+    /* Initialize ADC */
+    const adc_oneshot_unit_init_cfg_t init_config1 = {
+        .unit_id = BSP_ADC_UNIT,
+    };
+    BSP_ERROR_CHECK_RETURN_ERR(adc_oneshot_new_unit(&init_config1, &bsp_adc_handle));
+
+    return ESP_OK;
+}
+
+adc_oneshot_unit_handle_t bsp_adc_get_handle(void)
+{
+    return bsp_adc_handle;
+}
