+++ bsp/esp32_s3_eye/src/bsp_display.c
@@ -1,0 +1,300 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "esp_err.h"
+#include "esp_log.h"
+#include "esp_check.h"
+#include "bsp_err_check.h"
+#include "driver/ledc.h"
+#include "driver/spi_master.h"
+#include "esp_lcd_panel_io.h"
+#include "esp_lcd_panel_vendor.h"
+#include "esp_lcd_panel_ops.h"
+#include "bsp/display.h"
+
+#include "bsp/esp32_s3_eye.h"
+
+static const char *TAG = "ESP32-S3-EYE";
+#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+static lv_display_t *disp;
+#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+static bsp_lcd_handles_t disp_handles;
+
+// Bit number used to represent command and parameter
+#define LCD_CMD_BITS           8
+#define LCD_PARAM_BITS         8
+#define LCD_LEDC_CH            CONFIG_BSP_DISPLAY_BRIGHTNESS_LEDC_CH
+
+esp_err_t bsp_display_brightness_init(void)
+{
+    // Setup LEDC peripheral for PWM backlight control
+    const ledc_channel_config_t LCD_backlight_channel = {
+        .gpio_num = BSP_LCD_BACKLIGHT,
+        .speed_mode = LEDC_LOW_SPEED_MODE,
+        .channel = LCD_LEDC_CH,
+        .intr_type = LEDC_INTR_DISABLE,
+        .timer_sel = LEDC_TIMER_0,
+        .duty = 0,
+        .hpoint = 0
+    };
+    const ledc_timer_config_t LCD_backlight_timer = {
+        .speed_mode = LEDC_LOW_SPEED_MODE,
+        .duty_resolution = LEDC_TIMER_10_BIT,
+        .timer_num = LEDC_TIMER_0,
+        .freq_hz = 5000,
+        .clk_cfg = LEDC_AUTO_CLK
+    };
+
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_config(&LCD_backlight_timer));
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_channel_config(&LCD_backlight_channel));
+
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_brightness_deinit(void)
+{
+    const ledc_timer_config_t LCD_backlight_timer = {
+        .speed_mode = LEDC_LOW_SPEED_MODE,
+        .timer_num = LEDC_TIMER_0,
+        .deconfigure = 1
+    };
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_pause(LEDC_LOW_SPEED_MODE, LEDC_TIMER_0));
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_config(&LCD_backlight_timer));
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_brightness_set(int brightness_percent)
+{
+    if (brightness_percent > 100) {
+        brightness_percent = 100;
+    }
+    if (brightness_percent < 0) {
+        brightness_percent = 0;
+    }
+
+    ESP_LOGI(TAG, "Setting LCD backlight: %d%%", brightness_percent);
+    uint32_t duty_cycle = (1023 * (100 - brightness_percent)) / 100; // LEDC resolution set to 10bits, thus: 100% = 1023
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_set_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH, duty_cycle));
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_update_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH));
+
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_backlight_off(void)
+{
+    return bsp_display_brightness_set(0);
+}
+
+esp_err_t bsp_display_backlight_on(void)
+{
+    return bsp_display_brightness_set(100);
+}
+
+esp_err_t bsp_display_new(const bsp_display_config_t *config, esp_lcd_panel_handle_t *ret_panel,
+                          esp_lcd_panel_io_handle_t *ret_io)
+{
+    esp_err_t ret = ESP_OK;
+    bsp_lcd_handles_t handles;
+    ret = bsp_display_new_with_handles(config, &handles);
+
+    *ret_panel = handles.panel;
+    *ret_io = handles.io;
+
+    return ret;
+}
+
+esp_err_t bsp_display_new_with_handles(const bsp_display_config_t *config, bsp_lcd_handles_t *ret_handles)
+{
+    esp_err_t ret = ESP_OK;
+    assert(config != NULL && config->max_transfer_sz > 0);
+    ESP_RETURN_ON_ERROR(bsp_display_brightness_init(), TAG, "Brightness init failed");
+
+    /* DISPLAY - SPI */
+    ESP_LOGD(TAG, "Initialize SPI bus");
+    const spi_bus_config_t buscfg = {
+        .sclk_io_num = BSP_LCD_PCLK,
+        .mosi_io_num = BSP_LCD_DATA0,
+        .miso_io_num = GPIO_NUM_NC,
+        .quadwp_io_num = GPIO_NUM_NC,
+        .quadhd_io_num = GPIO_NUM_NC,
+
+        .max_transfer_sz = config->max_transfer_sz,
+    };
+    ESP_RETURN_ON_ERROR(spi_bus_initialize(BSP_LCD_SPI_NUM, &buscfg, SPI_DMA_CH_AUTO), TAG, "SPI init failed");
+
+    ESP_LOGD(TAG, "Install panel IO");
+    const esp_lcd_panel_io_spi_config_t io_config = {
+        .dc_gpio_num = BSP_LCD_DC,
+        .cs_gpio_num = BSP_LCD_CS,
+        .pclk_hz = BSP_LCD_PIXEL_CLOCK_HZ,
+        .lcd_cmd_bits = LCD_CMD_BITS,
+        .lcd_param_bits = LCD_PARAM_BITS,
+        .spi_mode = 0,
+        .trans_queue_depth = 10,
+    };
+    ESP_GOTO_ON_ERROR(esp_lcd_new_panel_io_spi((esp_lcd_spi_bus_handle_t)BSP_LCD_SPI_NUM, &io_config, &ret_handles->io),
+                      err, TAG, "New panel IO failed");
+    disp_handles.io = ret_handles->io;
+    ESP_LOGD(TAG, "Install LCD driver");
+    const esp_lcd_panel_dev_config_t panel_config = {
+        .reset_gpio_num = BSP_LCD_RST,
+        .flags.reset_active_high = 0,
+        .rgb_ele_order = BSP_LCD_COLOR_SPACE,
+        .bits_per_pixel = BSP_LCD_BITS_PER_PIXEL,
+    };
+    ESP_GOTO_ON_ERROR(esp_lcd_new_panel_st7789(ret_handles->io, &panel_config, &ret_handles->panel), err, TAG,
+                      "New panel failed");
+    disp_handles.panel = ret_handles->panel;
+
+    esp_lcd_panel_reset(ret_handles->panel);
+    esp_lcd_panel_init(ret_handles->panel);
+    esp_lcd_panel_invert_color(ret_handles->panel, true);
+    esp_lcd_panel_swap_xy(ret_handles->panel, false);
+    esp_lcd_panel_mirror(ret_handles->panel, false, false);
+
+    ESP_LOGI(TAG, "Display initialized with resolution %dx%d", BSP_LCD_H_RES, BSP_LCD_V_RES);
+    return ret;
+
+err:
+    bsp_display_delete();
+    return ret;
+}
+
+void bsp_display_delete(void)
+{
+    if (disp_handles.panel) {
+        esp_lcd_panel_del(disp_handles.panel);
+        disp_handles.panel = NULL;
+    }
+    if (disp_handles.io) {
+        esp_lcd_panel_io_del(disp_handles.io);
+        disp_handles.io = NULL;
+    }
+    spi_bus_free(BSP_LCD_SPI_NUM);
+
+    bsp_display_brightness_deinit();
+}
+
+#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+static lv_display_t *bsp_display_lcd_init(const bsp_display_cfg_t *cfg)
+{
+    assert(cfg != NULL);
+    esp_lcd_panel_io_handle_t io_handle = NULL;
+    const bsp_display_config_t bsp_disp_cfg = {
+        .max_transfer_sz = (BSP_LCD_H_RES * CONFIG_BSP_LCD_DRAW_BUF_HEIGHT) * sizeof(uint16_t),
+    };
+    BSP_ERROR_CHECK_RETURN_NULL(bsp_display_new(&bsp_disp_cfg, &disp_handles.panel, &io_handle));
+
+    esp_lcd_panel_disp_on_off(disp_handles.panel, true);
+
+    /* Add LCD screen */
+    ESP_LOGD(TAG, "Add LCD screen");
+    const lvgl_port_display_cfg_t disp_cfg = {
+        .io_handle = io_handle,
+        .panel_handle = disp_handles.panel,
+        .buffer_size = cfg->buffer_size,
+        .double_buffer = cfg->double_buffer,
+        .hres = BSP_LCD_H_RES,
+        .vres = BSP_LCD_V_RES,
+        .monochrome = false,
+        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
+        .rotation = {
+            .swap_xy = false,
+            .mirror_x = false,
+            .mirror_y = false,
+        },
+        .flags = {
+            .buff_dma = cfg->flags.buff_dma,
+            .buff_spiram = cfg->flags.buff_spiram,
+#if LVGL_VERSION_MAJOR >= 9
+            .swap_bytes = (BSP_LCD_BIGENDIAN ? true : false),
+#endif
+#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
+            .sw_rotate = false,                /* Avoid tearing is not supported for SW rotation */
+#else
+            .sw_rotate = cfg->flags.sw_rotate,
+#endif
+        }
+    };
+    return lvgl_port_add_disp(&disp_cfg);
+}
+#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+
+#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+lv_display_t *bsp_display_start(void)
+{
+    bsp_display_cfg_t cfg = {
+        .lvgl_port_cfg = ESP_LVGL_PORT_INIT_CONFIG(),
+        .buffer_size = BSP_LCD_H_RES * CONFIG_BSP_LCD_DRAW_BUF_HEIGHT,
+#if CONFIG_BSP_LCD_DRAW_BUF_DOUBLE
+        .double_buffer = 1,
+#else
+        .double_buffer = 0,
+#endif
+        .flags = {
+            .buff_dma = true,
+            .buff_spiram = false,
+            .sw_rotate = false,
+        }
+    };
+    return bsp_display_start_with_config(&cfg);
+}
+
+lv_display_t *bsp_display_start_with_config(const bsp_display_cfg_t *cfg)
+{
+    assert(cfg != NULL);
+    BSP_ERROR_CHECK_RETURN_NULL(lvgl_port_init(&cfg->lvgl_port_cfg));
+    BSP_NULL_CHECK(disp = bsp_display_lcd_init(cfg), NULL);
+
+    return disp;
+}
+
+lv_indev_t *bsp_display_get_input_dev(void)
+{
+    return NULL;
+}
+
+void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation)
+{
+    lv_disp_set_rotation(disp, rotation);
+}
+
+bool bsp_display_lock(uint32_t timeout_ms)
+{
+    return lvgl_port_lock(timeout_ms);
+}
+
+void bsp_display_unlock(void)
+{
+    lvgl_port_unlock();
+}
+
+static esp_err_t bsp_lcd_enter_sleep(void)
+{
+    assert(disp_handles.panel);
+    return esp_lcd_panel_disp_on_off(disp_handles.panel, false);
+}
+
+static esp_err_t bsp_lcd_exit_sleep(void)
+{
+    assert(disp_handles.panel);
+    return esp_lcd_panel_disp_on_off(disp_handles.panel, true);
+}
+
+esp_err_t bsp_display_enter_sleep(void)
+{
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_lcd_enter_sleep());
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_display_backlight_off());
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_exit_sleep(void)
+{
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_lcd_exit_sleep());
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_display_backlight_on());
+    return ESP_OK;
+}
+#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
