+++ docu/doxyfiles/templates/how_to_use/g12_camera.md
@@ -5,7 +5,7 @@ This function sets up the required I2C bus, video subsystem, and camera clock if
 
 ### Example Usage
 
-Camera usage can be quite complex. For a complete example, refer to the [`display_camera_csi`](https://github.com/espressif/esp-bsp/tree/master/examples/display_camera_csi) example in the BSP repository, or to the examples provided in the [`esp_video`](https://github.com/espressif/esp-video-components/tree/master/esp_video) component.
+Camera usage can be quite complex. For a complete example, refer to the [`display_camera_video`](https://github.com/espressif/esp-bsp/tree/master/examples/display_camera_video) example in the BSP repository, or to the examples provided in the [`esp_video`](https://github.com/espressif/esp-video-components/tree/master/esp_video) component.
 
 > [!NOTE]
 > Please, do not forget select right camera sensor in `menuconfig`
