+++ examples/display_camera_video/CMakeLists.txt
