+++ examples/display_camera_video/main/CMakeLists.txt
