+++ examples/display_camera_video/main/idf_component.yml
