+++ examples/display_camera_video/main/main.c
@@ -6,9 +6,9 @@
 
 /**
  * @file
- * @brief BSP Camera Example (MIPI-CSI)
- * @details Stream camera (MIPI-CSI) output to display (LVGL)
- * @example https://espressif.github.io/esp-launchpad/?flashConfigURL=https://espressif.github.io/esp-bsp/config.toml&app=display_camera-
+ * @brief BSP Camera Example
+ * @details Stream camera output to display (LVGL)
+ * @example https://espressif.github.io/esp-launchpad/?flashConfigURL=https://espressif.github.io/esp-bsp/config.toml&app=display_camera_video
  */
 
 #include <stdio.h>
@@ -21,7 +21,9 @@
 #include <unistd.h>
 #include <sys/ioctl.h>
 #include <linux/videodev2.h>
+#if SOC_PPA_SUPPORTED
 #include "driver/ppa.h"
+#endif
 #include "esp_private/esp_cache_private.h"
 #include "app_video.h"
 
@@ -34,12 +36,15 @@ typedef struct {
 } mmap_buf_t;
 
 static const char *TAG = "example";
+#if SOC_PPA_SUPPORTED
 static ppa_client_handle_t ppa_srm_handle = NULL;
+#endif
 static size_t data_cache_line_size = 0;
 static lv_obj_t *camera_canvas = NULL;
 static uint8_t *cam_buff[NUM_BUFS];
 static uint32_t cam_buff_size = 0;
 
+#if SOC_PPA_SUPPORTED
 static void app_ppa_init(void)
 {
     /* Initialize PPA */
@@ -109,12 +114,15 @@ static void calc_aspect_fit(
         *out_w = dst_h * src_aspect;
     }
 }
-
+#endif
 static void camera_video_frame_operation(uint8_t *camera_buf, uint8_t camera_buf_index, uint32_t camera_buf_hes,
         uint32_t camera_buf_ves, size_t camera_buf_len)
 {
+    uint32_t out_w = camera_buf_hes;
+    uint32_t out_h = camera_buf_ves ;
+    uint8_t *out_buf = camera_buf;
+#if SOC_PPA_SUPPORTED
     ppa_srm_rotation_angle_t rotation = PPA_SRM_ROTATION_ANGLE_0;
-    uint32_t out_w, out_h;
 
     switch (BSP_CAMERA_ROTATION) {
     case 0:
@@ -144,9 +152,11 @@ static void camera_video_frame_operation(uint8_t *camera_buf, uint8_t camera_buf
         cam_buff[camera_buf_index], out_w, out_h, cam_buff_size,
         rotation
     );
+    out_buf = cam_buff[camera_buf_index];
+#endif
 
     bsp_display_lock(0);
-    lv_canvas_set_buffer(camera_canvas, cam_buff[camera_buf_index], out_w, out_h, LV_COLOR_FORMAT_RGB565);
+    lv_canvas_set_buffer(camera_canvas, out_buf, out_w, out_h, LV_COLOR_FORMAT_RGB565);
     lv_obj_center(camera_canvas);
     lv_obj_invalidate(camera_canvas);
     bsp_display_unlock();
@@ -162,8 +172,10 @@ void app_main(void)
     /* Initialize Camera */
     bsp_camera_start(NULL);
 
+#if SOC_PPA_SUPPORTED
     /* Initialize PPA for scaling */
     app_ppa_init();
+#endif
 
     /* Get cache alignment */
     ret = esp_cache_get_alignment(MALLOC_CAP_SPIRAM, &data_cache_line_size);
