+++ examples/display_camera_video/sdkconfig.bsp.esp32_p4_eye
