+++ examples/display_camera_video/sdkconfig.bsp.esp32_s3_eye
@@ -4,25 +4,29 @@
 CONFIG_IDF_TARGET="esp32s3"
 CONFIG_BSP_SELECT_ESP32_S3_EYE=y
 
-CONFIG_ESPTOOLPY_FLASHMODE_QIO=y
 CONFIG_ESPTOOLPY_FLASHSIZE_4MB=y
-CONFIG_SPIRAM=y
-CONFIG_SPIRAM_SPEED_80M=y
-CONFIG_SPIRAM_MODE_OCT=y
 CONFIG_ESP_DEFAULT_CPU_FREQ_MHZ_240=y
 CONFIG_SPIRAM=y
-CONFIG_ESP32S3_SPIRAM_SUPPORT=y
 CONFIG_SPIRAM_MODE_OCT=y
 CONFIG_SPIRAM_SPEED_80M=y
-CONFIG_SPIRAM_TRY_ALLOCATE_WIFI_LWIP=y
-CONFIG_SPIRAM_MALLOC_RESERVE_INTERNAL=8192
-CONFIG_SPIRAM_ALLOW_BSS_SEG_EXTERNAL_MEMORY=y
 CONFIG_CODEC_I2C_BACKWARD_COMPATIBLE=n
+CONFIG_ESPTOOLPY_FLASHFREQ_80M=y
+CONFIG_ESPTOOLPY_FLASHMODE_QIO=y
+
+
+# Camera config
+CONFIG_FREERTOS_HZ=1000
+CONFIG_CAMERA_XCLK_USE_LEDC=y
+CONFIG_EXAMPLE_ENABLE_DVP_CAM_SENSOR=y
+CONFIG_CAMERA_OV2640=y
+CONFIG_CAMERA_OV2640_DVP_RGB565_240X240_25FPS=y
+CONFIG_CAM_CTRL_SPI_ENABLE=n
+
 
 ## LVGL9 ##
 CONFIG_LV_CONF_SKIP=y
 
-#CLIB default
+# CLIB default
 CONFIG_LV_USE_CLIB_MALLOC=y
 CONFIG_LV_USE_CLIB_SPRINTF=y
 CONFIG_LV_USE_CLIB_STRING=y
@@ -31,10 +35,3 @@ CONFIG_LV_USE_CLIB_STRING=y
 CONFIG_LV_USE_OBSERVER=y
 CONFIG_LV_USE_SYSMON=y
 CONFIG_LV_USE_PERF_MONITOR=y
-
-## LVGL8 - uncomment, when using LVGL8 ##
-# CONFIG_LV_USE_PERF_MONITOR=y
-# CONFIG_LV_COLOR_16_SWAP=y
-# CONFIG_LV_MEM_CUSTOM=y
-# CONFIG_LV_MEMCPY_MEMSET_STD=y
-# CONFIG_LV_SPRINTF_CUSTOM=y
