+++ examples/display_camera_video/sdkconfig.bsp.m5stack_tab5
