+++ examples/display_camera_video/sdkconfig.defaults
