+++ examples/display_lvgl_benchmark/main/main.c
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2025 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2025-2026 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: CC0-1.0
  */
@@ -31,30 +31,6 @@ void app_main(void)
     };
     cfg.lvgl_port_cfg.task_stack = 10000;
     bsp_display_start_with_config(&cfg);
-#elif defined(BSP_BOARD_ESP32_S3_EYE)
-    /* Only for esp32_s3_eye */
-    bsp_display_cfg_t cfg = {
-        .lvgl_port_cfg = {
-            .task_priority = CONFIG_BSP_DISPLAY_LVGL_TASK_PRIORITY,
-            .task_stack = 10000,
-            .task_affinity = 1,
-            .timer_period_ms = CONFIG_BSP_DISPLAY_LVGL_TICK,
-            .task_max_sleep_ms = CONFIG_BSP_DISPLAY_LVGL_MAX_SLEEP,
-        },
-        .buffer_size = BSP_LCD_DRAW_BUFF_SIZE,
-        .double_buffer = BSP_LCD_DRAW_BUFF_DOUBLE,
-        .flags = {
-#if (CONFIG_BSP_DISPLAY_LVGL_BUFFER_IN_PSRAM && CONFIG_SPIRAM)
-            .buff_dma = false,
-            .buff_spiram = true,
-#else
-            .buff_dma = true,
-            .buff_spiram = false,
-#endif
-        }
-    };
-    bsp_display_start_with_config(&cfg);
-#else
     bsp_display_start();
 #endif
